Add PR labeler and release drafter config (#1)
* Add PR labeler and release drafter config

Previously we set release information manually. Now PRs are categorized
and added in their respective sections depending on branch name
they originated from.

For more details on how we categorize, see PR-labeler configuration:
[[./.github/workflows/config/pr-labeler.yaml]]

* Add spec for tag parsing for check cmd

Previously the spec wasn't available for tag parsing. Because an
interface and a custom Var is used for tag parsing in cobra, it
makes sense to add the spec.

* Package binary release as gzip

Previously we were only uploading a binary, but it turns out that
for some internal automation it makes sense to also upload a gzip file.